move: email 도메인에 적합. global에서 domain 하위 패키지로 이동
